feat: remove CF-specific logic
