+++ packages/playground/server-components/package.json
@@ -9,7 +9,7 @@
     "build": "yarn build:client && yarn build:server && yarn build:worker",
     "build:client": "vite build --outDir dist/client --manifest",
     "build:server": "vite build --outDir dist/server --ssr @shopify/hydrogen/platforms/node",
-    "build:worker": "cross-env WORKER=true vite build --outDir dist/worker --ssr @shopify/hydrogen/platforms/cfw-event"
+    "build:worker": "cross-env WORKER=true vite build --outDir dist/worker --ssr worker"
   },
   "devDependencies": {
     "cross-env": "^7.0.3",
